// src/types.ts
export interface AnnouncementItem {
  id: string;
  title: string;
  content: string;
  date: string;
  time: string;
  speaker: string;
  category: 'kajian' | 'jumat' | 'tarbiyah' | 'sosial' | 'keuangan';
  isActive: boolean;
}